+++ integrations/github/tests/github/core/exporters/test_repository_exporter.py
@@ -5,11 +5,13 @@ import httpx
 from github.core.exporters.repository_exporter import (
     RestRepositoryExporter,
 )
+from pydantic import ValidationError
 from github.core.options import ListRepositoryOptions, SingleRepositoryOptions
 from integration import GithubPortAppConfig
 from port_ocean.context.event import event_context
 from github.helpers.models import RepoSearchParams
 from github.clients.http.rest_client import GithubRestClient
+from integration import GithubRepositorySelector
 
 
 TEST_REPOS = [
@@ -443,3 +445,55 @@ class TestRestRepositoryExporter:
             )
 
             assert result["custom_properties"] == {}
+
+
+
+def test_normalized_relations_from_included_relations_alias() -> None:
+    selector = GithubRepositorySelector.parse_obj(
+        {
+            "query": "true",
+            "includedRelations": {
+                "teams": True,
+                "sbom": False,
+                "pages": True,
+                "collaborators": {"affiliation": "direct"},
+            },
+        }
+    )
+
+    assert selector.normalized_relations == {
+        "teams": {"include": True},
+        "pages": {"include": True},
+        "collaborators": {"include": True, "affiliation": "direct"},
+    }
+
+
+def test_included_relations_cannot_be_supplied_with_include() -> None:
+    with pytest.raises(ValidationError) as exc:
+        GithubRepositorySelector.parse_obj(
+            {
+                "query": "true",
+                "include": ["teams"],
+                "includedRelations": {"collaborators": {"affiliation": "all"}},
+            }
+        )
+
+    assert "You cannot supply both 'include' and 'includedRelations'" in str(exc.value)
+
+
+def test_normalized_relations_falls_back_to_include_list() -> None:
+    selector = GithubRepositorySelector.parse_obj(
+        {"query": "true", "include": ["teams", "sbom"]}
+    )
+
+    assert selector.normalized_relations == {
+        "teams": {"include": True},
+        "sbom": {"include": True},
+    }
+
+
+def test_included_relations_forbids_unknown_keys() -> None:
+    with pytest.raises(ValidationError):
+        GithubRepositorySelector.parse_obj(
+            {"query": "true", "includedRelations": {"unknown": True}}
+        )
